refactor: replace types.ts from components
